+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/errorprone/AvoidDuplicateLiteralsRule.java
@@ -17,6 +17,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.logging.Logger;
 
 import org.apache.commons.lang3.StringUtils;
 
@@ -34,6 +35,7 @@ import net.sourceforge.pmd.properties.StringProperty;
 
 
 public class AvoidDuplicateLiteralsRule extends AbstractJavaRule {
+    private static final Logger LOG = Logger.getLogger(AvoidDuplicateLiteralsRule.class.getName());
 
     public static final PropertyDescriptor<Integer> THRESHOLD_DESCRIPTOR
             = PropertyFactory.intProperty("maxDuplicateLiterals")
@@ -45,16 +47,34 @@ public class AvoidDuplicateLiteralsRule extends AbstractJavaRule {
     public static final BooleanProperty SKIP_ANNOTATIONS_DESCRIPTOR = new BooleanProperty("skipAnnotations",
             "Skip literals within annotations", false, 2.0f);
 
-    public static final StringProperty EXCEPTION_LIST_DESCRIPTOR = new StringProperty("exceptionList",
-            "Strings to ignore", null, 3.0f);
+    // This set of properties is impossible to convert to the new framework before 7.0.0
+    // It looks like it tried to implement itself delimiter escaping and such, which we
+    // can't emulate without significant effort (wasted effort, bc delimiters will anyway
+    // be scrapped).
+    // TODO 7.0.0:
+    // EXCEPTION_LIST_DESCRIPTOR -> PropertyDescriptor<Set<String>>
+    // delete SEPARATOR_DESCRIPTOR, EXCEPTION_FILE_DESCRIPTOR
+    // Try hard to convert the properties to the seq syntax, using the separator descriptor
+
+    // TODO We use the old builders here, bc of the null default value
+    public static final StringProperty EXCEPTION_LIST_DESCRIPTOR
+            = StringProperty.named("exceptionList")
+                            .desc("List of literals to ignore. "
+                                          + "A literal is ignored if its image can be found in this list. "
+                                          + "Components of this list should not be surrounded by double quotes.")
+                            .defaultValue(null)
+                            .build();
 
     public static final CharacterProperty SEPARATOR_DESCRIPTOR = new CharacterProperty("separator",
             "Ignore list separator", ',', 4.0f);
 
+    @Deprecated
     public static final FileProperty EXCEPTION_FILE_DESCRIPTOR = new FileProperty("exceptionfile",
-            "File containing strings to skip (one string per line), only used if ignore list is not set. "
+            "deprecated!(Use 'exceptionList' property) File containing strings to skip (one string per line), only used if ignore list is not set. "
             + "File must be UTF-8 encoded.", null, 5.0f);
 
+    /** @deprecated This ad-hoc solution will be integrated into the global properties framework somehow */
+    @Deprecated
     public static class ExceptionParser {
 
         private static final char ESCAPE_CHAR = '\\';
@@ -115,10 +135,18 @@ public class AvoidDuplicateLiteralsRule extends AbstractJavaRule {
         literals.clear();
 
         if (getProperty(EXCEPTION_LIST_DESCRIPTOR) != null) {
+            if (isPropertyOverridden(SEPARATOR_DESCRIPTOR)) {
+                LOG.warning("Rule AvoidDuplicateLiterals uses deprecated property 'separator'. "
+                                    + "Future versions of PMD will remove support for this property. "
+                                    + "Please use the default separator (',') and avoid setting this property instead.");
+            }
             ExceptionParser p = new ExceptionParser(getProperty(SEPARATOR_DESCRIPTOR));
             exceptions = p.parse(getProperty(EXCEPTION_LIST_DESCRIPTOR));
         } else if (getProperty(EXCEPTION_FILE_DESCRIPTOR) != null) {
             exceptions = new HashSet<>();
+            LOG.warning("Rule AvoidDuplicateLiterals uses deprecated property 'exceptionFile'. "
+                                + "Future versions of PMD will remove support for this property. "
+                                + "Please use 'exceptionList' instead.");
             try (LineNumberReader reader = getLineReader()) {
                 String line;
                 while ((line = reader.readLine()) != null) {
@@ -147,8 +175,7 @@ public class AvoidDuplicateLiteralsRule extends AbstractJavaRule {
             if (occurrences.size() >= threshold) {
                 ASTLiteral first = occurrences.get(0);
                 String rawImage = first.getEscapedStringLiteral();
-                Object[] args = new Object[] { rawImage, Integer.valueOf(occurrences.size()),
-                    Integer.valueOf(first.getBeginLine()), };
+                Object[] args = {rawImage, occurrences.size(), first.getBeginLine(), };
                 addViolation(data, first, args);
             }
         }